Revert "Reports email pagination"
